Fixed alignment bugs relating to room transitions and direction changes.
Because the hitbox is different depending on which way the player faces,
you could effectively get through a barrier by facing right and jumping
while right below a wall, then turning left in midair.  Corrected this
by using an offset.  Switching between rooms had a similar issue because
placement is based on the player's rect while wall collisions are
determined by the hitbox, so used offset for placement in the new room.